Changed the Divide-method so that it returns float instead of int

+++ src/main/java/com/example/springdocker/MyMathCalculator.java
@@ -17,7 +17,7 @@ public class MyMathCalculator {
         return a * b;
     }
 
-    public int divide(int a, int b) {
+    public float divide(float a, float b) {
         return a / b;
     }
 }
